Refactor the `ImageFormat` trait out of existence
