+++ espflash/src/image_format.rs
@@ -1,22 +1,104 @@
+//! ESP-IDF application binary image format
+
 use std::{borrow::Cow, io::Write, iter::once, mem::size_of};
 
-use bytemuck::{bytes_of, from_bytes};
+use bytemuck::{bytes_of, from_bytes, Pod, Zeroable};
 use esp_idf_part::{Partition, PartitionTable, Type};
 use sha2::{Digest, Sha256};
 
 use crate::{
     elf::{CodeSegment, FirmwareImage, RomSegment},
     error::Error,
-    flasher::FlashSettings,
-    image_format::{
-        update_checksum, ImageFormat, ImageHeader, SegmentHeader, ESP_CHECKSUM_MAGIC, ESP_MAGIC,
-        WP_PIN_DISABLED,
-    },
+    flasher::{FlashFrequency, FlashMode, FlashSettings, FlashSize},
     targets::{Chip, Esp32Params},
 };
 
+const ESP_CHECKSUM_MAGIC: u8 = 0xef;
+const ESP_MAGIC: u8 = 0xE9;
 const IROM_ALIGN: u32 = 0x10000;
 const SEG_HEADER_LEN: u32 = 8;
+const WP_PIN_DISABLED: u8 = 0xEE;
+
+/// Firmware header used by the ESP-IDF bootloader.
+///
+/// ## Header documentation:
+/// * [Header](https://docs.espressif.com/projects/esptool/en/latest/esp32c3/advanced-topics/firmware-image-format.html#file-header)
+/// * [Extended header](https://docs.espressif.com/projects/esptool/en/latest/esp32c3/advanced-topics/firmware-image-format.html#extended-file-header)
+#[derive(Debug, Clone, Copy, Pod, Zeroable)]
+#[repr(C, packed)]
+#[doc(alias = "esp_image_header_t")]
+struct ImageHeader {
+    magic: u8,
+    segment_count: u8,
+    /// Flash read mode (esp_image_spi_mode_t)
+    flash_mode: u8,
+    /// ..4 bits are flash chip size (esp_image_flash_size_t)
+    /// 4.. bits are flash frequency (esp_image_spi_freq_t)
+    #[doc(alias = "spi_size")]
+    #[doc(alias = "spi_speed")]
+    flash_config: u8,
+    entry: u32,
+
+    // extended header part
+    wp_pin: u8,
+    clk_q_drv: u8,
+    d_cs_drv: u8,
+    gd_wp_drv: u8,
+    chip_id: u16,
+    min_rev: u8,
+    /// Minimum chip revision supported by image, in format: major * 100 + minor
+    min_chip_rev_full: u16,
+    /// Maximal chip revision supported by image, in format: major * 100 + minor
+    max_chip_rev_full: u16,
+    reserved: [u8; 4],
+    append_digest: u8,
+}
+
+impl Default for ImageHeader {
+    fn default() -> Self {
+        Self {
+            magic: ESP_MAGIC,
+            segment_count: 3,
+            flash_mode: FlashMode::default() as _,
+            flash_config: ((FlashSize::default() as u8) << 4) | FlashFrequency::default() as u8,
+            entry: 0,
+            wp_pin: WP_PIN_DISABLED,
+            clk_q_drv: 0,
+            d_cs_drv: 0,
+            gd_wp_drv: 0,
+            chip_id: Default::default(),
+            min_rev: 0,
+            min_chip_rev_full: 0,
+            max_chip_rev_full: u16::MAX,
+            reserved: Default::default(),
+            append_digest: 1,
+        }
+    }
+}
+
+impl ImageHeader {
+    /// Updates flash size and speed filed.
+    pub fn write_flash_config(
+        &mut self,
+        size: FlashSize,
+        freq: FlashFrequency,
+        chip: Chip,
+    ) -> Result<(), Error> {
+        let flash_size = size.encode_flash_size()?;
+        let flash_speed = freq.encode_flash_frequency(chip)?;
+
+        // bit field
+        self.flash_config = (flash_size << 4) | flash_speed;
+        Ok(())
+    }
+}
+
+#[derive(Debug, Clone, Copy, Pod, Zeroable)]
+#[repr(C, packed)]
+struct SegmentHeader {
+    addr: u32,
+    length: u32,
+}
 
 /// Image format for ESP32 family chips using the second-stage bootloader from
 /// ESP-IDF
@@ -188,8 +270,9 @@ impl<'a> IdfBootloaderFormat<'a> {
             data: Cow::Owned(data),
         };
 
-        // If the user did not specify a partition offset, we need to assume that the partition
-        // offset is (first partition offset) - 0x1000, since this is the most common case.
+        // If the user did not specify a partition offset, we need to assume that the
+        // partition offset is (first partition offset) - 0x1000, since this is
+        // the most common case.
         let partition_table_offset = partition_table_offset.unwrap_or_else(|| {
             let partitions = partition_table.partitions();
             let first_partition = partitions
@@ -209,10 +292,8 @@ impl<'a> IdfBootloaderFormat<'a> {
             partition_table_offset,
         })
     }
-}
 
-impl<'a> ImageFormat<'a> for IdfBootloaderFormat<'a> {
-    fn flash_segments<'b>(&'b self) -> Box<dyn Iterator<Item = RomSegment<'b>> + 'b>
+    pub fn flash_segments<'b>(&'b self) -> Box<dyn Iterator<Item = RomSegment<'b>> + 'b>
     where
         'a: 'b,
     {
@@ -252,18 +333,18 @@ impl<'a> ImageFormat<'a> for IdfBootloaderFormat<'a> {
         )
     }
 
-    fn ota_segments<'b>(&'b self) -> Box<dyn Iterator<Item = RomSegment<'b>> + 'b>
+    pub fn ota_segments<'b>(&'b self) -> Box<dyn Iterator<Item = RomSegment<'b>> + 'b>
     where
         'a: 'b,
     {
         Box::new(once(self.flash_segment.borrow()))
     }
 
-    fn app_size(&self) -> u32 {
+    pub fn app_size(&self) -> u32 {
         self.app_size
     }
 
-    fn part_size(&self) -> Option<u32> {
+    pub fn part_size(&self) -> Option<u32> {
         Some(self.part_size)
     }
 }
@@ -346,47 +427,30 @@ fn save_segment(data: &mut Vec<u8>, segment: &CodeSegment, checksum: u8) -> Resu
     Ok(update_checksum(segment.data(), checksum))
 }
 
+/// Update the checksum with the given data
+fn update_checksum(data: &[u8], mut checksum: u8) -> u8 {
+    for byte in data {
+        checksum ^= *byte;
+    }
+
+    checksum
+}
+
 #[cfg(test)]
-pub mod tests {
-    use std::fs;
-
+mod tests {
     use super::*;
-    use crate::{elf::ElfFirmwareImage, image_format::FlashFrequency};
-
-    // Copied from: src/targets/esp32.rs
-    const PARAMS: Esp32Params = Esp32Params::new(
-        0x1000,
-        0x1_0000,
-        0x3f_0000,
-        0,
-        FlashFrequency::_40Mhz,
-        include_bytes!("../../resources/bootloaders/esp32-bootloader.bin"),
-    );
 
     #[test]
-    fn test_idf_bootloader_format() {
-        let input_bytes = fs::read("tests/resources/esp32_hal_blinky").unwrap();
-        let expected_bin = fs::read("tests/resources/esp32_hal_blinky.bin").unwrap();
+    fn test_flash_config_write() {
+        let mut header = ImageHeader::default();
+        header
+            .write_flash_config(FlashSize::_4Mb, FlashFrequency::_40Mhz, Chip::Esp32c3)
+            .unwrap();
+        assert_eq!(header.flash_config, 0x20);
 
-        let image = ElfFirmwareImage::try_from(input_bytes.as_slice()).unwrap();
-        let flash_image = IdfBootloaderFormat::new(
-            &image,
-            Chip::Esp32,
-            0,
-            PARAMS,
-            None,
-            None,
-            None,
-            None,
-            FlashSettings::default(),
-        )
-        .unwrap();
-
-        let segments = flash_image.flash_segments().collect::<Vec<_>>();
-        assert_eq!(segments.len(), 3);
-
-        let buf = segments[2].data.as_ref();
-        assert_eq!(expected_bin.len(), buf.len());
-        assert_eq!(expected_bin.as_slice(), buf);
+        header
+            .write_flash_config(FlashSize::_32Mb, FlashFrequency::_80Mhz, Chip::Esp32s3)
+            .unwrap();
+        assert_eq!(header.flash_config, 0x5F);
     }
 }
