--- espflash/src/image_format/mod.rs
@@ -1,149 +1,0 @@
-//! ESP-IDF application binary image format
-
-use bytemuck::{Pod, Zeroable};
-
-pub use self::idf_bootloader::IdfBootloaderFormat;
-use crate::{
-    elf::RomSegment,
-    error::Error,
-    flasher::{FlashFrequency, FlashMode, FlashSize},
-    targets::Chip,
-};
-
-mod idf_bootloader;
-
-const ESP_CHECKSUM_MAGIC: u8 = 0xef;
-const ESP_MAGIC: u8 = 0xE9;
-const WP_PIN_DISABLED: u8 = 0xEE;
-
-/// Firmware header used by the ESP-IDF bootloader.
-///
-/// ## Header documentation:
-/// * [Header](https://docs.espressif.com/projects/esptool/en/latest/esp32c3/advanced-topics/firmware-image-format.html#file-header)
-/// * [Extended header](https://docs.espressif.com/projects/esptool/en/latest/esp32c3/advanced-topics/firmware-image-format.html#extended-file-header)
-#[derive(Debug, Clone, Copy, Pod, Zeroable)]
-#[repr(C, packed)]
-#[doc(alias = "esp_image_header_t")]
-struct ImageHeader {
-    magic: u8,
-    segment_count: u8,
-    /// Flash read mode (esp_image_spi_mode_t)
-    flash_mode: u8,
-    /// ..4 bits are flash chip size (esp_image_flash_size_t)
-    /// 4.. bits are flash frequency (esp_image_spi_freq_t)
-    #[doc(alias = "spi_size")]
-    #[doc(alias = "spi_speed")]
-    flash_config: u8,
-    entry: u32,
-
-    // extended header part
-    wp_pin: u8,
-    clk_q_drv: u8,
-    d_cs_drv: u8,
-    gd_wp_drv: u8,
-    chip_id: u16,
-    min_rev: u8,
-    /// Minimum chip revision supported by image, in format: major * 100 + minor
-    min_chip_rev_full: u16,
-    /// Maximal chip revision supported by image, in format: major * 100 + minor
-    max_chip_rev_full: u16,
-    reserved: [u8; 4],
-    append_digest: u8,
-}
-
-impl Default for ImageHeader {
-    fn default() -> Self {
-        Self {
-            magic: ESP_MAGIC,
-            segment_count: 3,
-            flash_mode: FlashMode::default() as _,
-            flash_config: ((FlashSize::default() as u8) << 4) | FlashFrequency::default() as u8,
-            entry: 0,
-            wp_pin: WP_PIN_DISABLED,
-            clk_q_drv: 0,
-            d_cs_drv: 0,
-            gd_wp_drv: 0,
-            chip_id: Default::default(),
-            min_rev: 0,
-            min_chip_rev_full: 0,
-            max_chip_rev_full: u16::MAX,
-            reserved: Default::default(),
-            append_digest: 1,
-        }
-    }
-}
-
-impl ImageHeader {
-    /// Updates flash size and speed filed.
-    pub fn write_flash_config(
-        &mut self,
-        size: FlashSize,
-        freq: FlashFrequency,
-        chip: Chip,
-    ) -> Result<(), Error> {
-        let flash_size = size.encode_flash_size()?;
-        let flash_speed = freq.encode_flash_frequency(chip)?;
-
-        // bit field
-        self.flash_config = (flash_size << 4) | flash_speed;
-        Ok(())
-    }
-}
-
-#[derive(Debug, Clone, Copy, Pod, Zeroable)]
-#[repr(C, packed)]
-struct SegmentHeader {
-    addr: u32,
-    length: u32,
-}
-
-/// Operations for working with firmware image formats
-pub trait ImageFormat<'a>: Send {
-    /// Get the ROM segments needed when flashing to device
-    fn flash_segments<'b>(&'b self) -> Box<dyn Iterator<Item = RomSegment<'b>> + 'b>
-    where
-        'a: 'b;
-
-    /// Get the ROM segments to save when exporting for OTA
-    ///
-    /// Compared to `flash_segments` this excludes things like bootloader and
-    /// partition table
-    fn ota_segments<'b>(&'b self) -> Box<dyn Iterator<Item = RomSegment<'b>> + 'b>
-    where
-        'a: 'b;
-
-    /// The size of the application binary
-    fn app_size(&self) -> u32;
-
-    /// If applicable, the size of the application partition (if it can be
-    /// determined)
-    fn part_size(&self) -> Option<u32>;
-}
-
-/// Update the checksum with the given data
-fn update_checksum(data: &[u8], mut checksum: u8) -> u8 {
-    for byte in data {
-        checksum ^= *byte;
-    }
-
-    checksum
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_flash_config_write() {
-        let mut header = ImageHeader::default();
-        header
-            .write_flash_config(FlashSize::_4Mb, FlashFrequency::_40Mhz, Chip::Esp32c3)
-            .unwrap();
-        assert_eq!(header.flash_config, 0x20);
-
-        header
-            .write_flash_config(FlashSize::_32Mb, FlashFrequency::_80Mhz, Chip::Esp32s3)
-            .unwrap();
-        assert_eq!(header.flash_config, 0x5F);
-    }
-}
