+++ espflash/src/targets/esp32p4.rs
@@ -6,7 +6,7 @@ use crate::{
     elf::FirmwareImage,
     error::Error,
     flasher::{FlashData, FlashFrequency},
-    image_format::{IdfBootloaderFormat, ImageFormat},
+    image_format::IdfBootloaderFormat,
     targets::{Chip, Esp32Params, ReadEFuse, SpiRegisters, Target, XtalFrequency},
 };
 
@@ -75,7 +75,7 @@ impl Target for Esp32p4 {
         flash_data: FlashData,
         _chip_revision: Option<(u32, u32)>,
         xtal_freq: XtalFrequency,
-    ) -> Result<Box<dyn ImageFormat<'a> + 'a>, Error> {
+    ) -> Result<IdfBootloaderFormat<'a>, Error> {
         if xtal_freq != XtalFrequency::_40Mhz {
             return Err(Error::UnsupportedFeature {
                 chip: Chip::Esp32p4,
@@ -83,7 +83,7 @@ impl Target for Esp32p4 {
             });
         }
 
-        Ok(Box::new(IdfBootloaderFormat::new(
+        IdfBootloaderFormat::new(
             image,
             Chip::Esp32p4,
             flash_data.min_chip_rev,
@@ -93,7 +93,7 @@ impl Target for Esp32p4 {
             flash_data.target_app_partition,
             flash_data.bootloader,
             flash_data.flash_settings,
-        )?))
+        )
     }
 
     fn spi_registers(&self) -> SpiRegisters {
